Allow custom clocks by moving the chrono_controller into it's header. #define NUCLEAR_CUSTOM_CLOCK to provide a stub that must be implemented in the main codebase and linked
